+++ app/components/storyboard/drawer/drawer-contents.js
@@ -1,5 +1,5 @@
+import { AuthContext } from '@/app/_layout';
 import DropdownInnerOption from '@/app/components/storyboard/dropdown-inner-option';
-import ChatStorageData from '@/app/constants/storyboard/CHATTING_STORAGE.js';
 import StoryboardStorageData from '@/app/constants/storyboard/STORYBOARD_STORAGE.js';
 import useThemedStyle from '@/app/hooks/use-themed-style';
 import ChatStorageIcon from '@/assets/svgs/storyboard/archive.js';
@@ -7,7 +7,7 @@ import NewChatIcon from '@/assets/svgs/storyboard/chat.js';
 import NewStoryboardIcon from '@/assets/svgs/storyboard/new-storyboard';
 import StoryboardIcon from '@/assets/svgs/storyboard/storyboard.js';
 import { useRouter } from 'expo-router';
-import { useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { Pressable, Text, View } from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
@@ -78,13 +78,61 @@ export default function StoryboardDrawerDefault({ onClose }) {
     );
   };
 
+  // 채팅 데이터
+  const { accessToken } = useContext(AuthContext);
+
+  const [myChatList, setMyChatList] = useState([]);
+
+  useEffect(() => {
+    fetch('https://dev.crezipsa.site/api/chats', {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    })
+      .then(async res => {
+        const raw = await res.text();
+        console.log('my chats api status', res.status);
+        console.log('my chats api raw response', raw);
+
+        let data = null;
+        try {
+          data = raw ? JSON.parse(raw) : null;
+        } catch (e) {
+          console.error('Failed to parse JSON response', e);
+        }
+
+        if (!res.ok) throw new Error(`My Chats API error: ${res.status}`);
+        return data;
+      })
+      .then(data => {
+        const result = data?.result;
+        if (!result || !Array.isArray(result)) {
+          console.error('Invalid my chats data format', data);
+          return;
+        }
+        const nextList = result
+          .filter(item => item?.lastMessageAt)
+          .sort(
+            (a, b) =>
+              new Date(b.lastMessageAt).getTime() -
+              new Date(a.lastMessageAt).getTime(),
+          );
+
+        setMyChatList(nextList.slice(0, 5));
+      })
+      .catch(error => {
+        console.error('Failed to fetch my chats data', error);
+      });
+  }, [accessToken]);
+
   return (
     <View style={{ gap: 16 }}>
       <Pressable
         style={styles.drawerOption}
         onPress={() => {
           onClose();
-          // 채팅 초기화
+          router.replace('/storyboard');
         }}
       >
         <NewChatIcon color={primaryColors.color} size={20} />
@@ -94,7 +142,7 @@ export default function StoryboardDrawerDefault({ onClose }) {
         style={styles.drawerOption}
         onPress={() => {
           onClose();
-          // 스토리보드 생성
+          // 라우팅 추가
         }}
       >
         <NewStoryboardIcon color={primaryColors.color} size={20} />
@@ -124,12 +172,13 @@ export default function StoryboardDrawerDefault({ onClose }) {
       {/* 채팅 드롭다운 */}
       {isChatDropdownVisible && (
         <View style={styles.dropdownContainer}>
-          {ChatStorageData.slice(0, 5).map((item, index) => (
+          {myChatList.slice(0, 5).map(item => (
             <DropdownInnerOption
-              key={item.id}
+              key={item.chatRoomId}
               label="chatting"
-              text={item.text}
-              route={item.route}
+              text={item.title}
+              route={item.chatRoomId}
+              chatRoomId={item.chatRoomId}
               onClose={onClose}
             />
           ))}
